Moved draw rules function to settings drawer
